+++ CHANGES.rst
@@ -3,12 +3,20 @@ Changelog
 
 There's a frood who really knows where his towel is.
 
-2.0.3 (unreleased)
-^^^^^^^^^^^^^^^^^^
+2.1 (unreleased)
+^^^^^^^^^^^^^^^^
+
+- Removed deprecated portal_actionicons registration
+  [keul]
+
+- Translation fixes: some label were not translated
+  [keul]
+
+- Facebook "Share" button now provided. See `#15`_.
+  [keul]
 
 - Added missing migration step for refreshing CSS registry
   [keul]
-
 
 2.0.2 (2015-02-23)
 ^^^^^^^^^^^^^^^^^^
@@ -214,3 +222,4 @@ There's a frood who really knows where his towel is.
 
 .. _`#36`: https://github.com/collective/sc.social.like/issues/36
 .. _`#38`: https://github.com/collective/sc.social.like/issues/38
+.. _`#15`: https://github.com/collective/sc.social.like/pull/15